interception_core_handler/requirements.txt
pathlib >= 1.0.1
requests >= 2.18.4
httplib2 >= 0.9.2
kafka-python
IPy
python-logstash
logstashhandler